one global-requirements.txt to rule them all
... and in the darkness bind them.

This moves the contents of requirements.txt and
test-requirements.txt to a single file, global-requirements.txt.
This means that regardless of whether a requirement is in
either file it will get checked. This massively simplifies the
checking and upgrading of req and test-req files.

This also adds in a unit testing framework for the update.py
script to ensure it does what we think it does. This is
accomplished by making a fake tree with a set of req and test-req
files, running update.py, and ensuring the updates were made
that we expected. It includes testing for the oslo url case,
as well as test-requires.

(cherry picked from commit ec9cf616904d92509b8b44662e7097352d894484)
(cherry picked from commit f7fbf480daf8dbb09650a6de4f21c1a487bfdedb)

Change-Id: Ib9b86ade4cb8317509e218aec31f32e5d08f4035