Update patch set 1
Patch Set 1:

This is a nice example to consider properly

It looks like paramiko 1.8.0 released in Oct 2012 fixed the bug. Can we make it

  paramiko>=1.8.0

instead?

This is the fix AFAICT: https://github.com/paramiko/paramiko/commits/f8239015

1.7.6 was released in November 2009

We don't want a newer version for any new APIs, purely for this bugfix

That's a really easily backportable fix, so I expect some of the more conservative distros might consider backporting the fix rather than rebasing to a much more recent version

That said ... this metadata is primarly for people installing deps from pypi and for our test infrastructure and in both cases >=1.7.6 is innacurate

I'd be fine with >=1.8.0

Patch-set: 1
Label: Code-Review=0
Label: Workflow=0