Add testresources and testscenarios.
Now that we're on testtools/testr in places, we can start using
testresources and testscenarios to optimize our test suites a bit.

testresources allows for declarative fixture creation in such a way that
an optimizing test suite can group tests that need an expensive-to-set-up
fixture together and use a reset method on the fixture between tests.

testscenarios allows for declarations of combinations of inputs and outputs
to a single tests, which can greatly simplify large amounts of repetitive
code that's trying to grab coverage of option sets.

Change-Id: I834813071c519086f06f80bf418c949bbc47c0ab