Update patch set 1
Patch Set 1:

Correct, the problem of a broken colorizer affects Debian, but we do not need to wait for Wheezy to fix this as Thomas asked. This problem is independent of distro packaging. (We need to make sure we support both people using pypi and debian packages by ensuring the colorizer is subunit v1 and v2 capable).

Patch-set: 1