Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

The check test is failing because it is currently hardcoded to look for pip-requires and test-requires. If folks are happy with this change and want it to pass the gate I will update the test.

Patch-set: 1
Label: Code-Review=+1
Label: Workflow=0