Update patch set 2
Patch Set 2:

part of the reason for requirements though was to give the distros a chance to weigh in and say "no way, that's insane to package that".

So I think you can +2 something if it's already in the distros. And if not, we really need both a Canonical and Red Hat person at least so say "sure, we can probably handle that".

Patch-set: 2
Label: Code-Review=0