Update patch set 1
Patch Set 1:

I think this script was documented somewhere as being recommended to use, and it works just fine except for the bug in it.

Why the -1 when the code looks good? Should I delete the code instead? I think this special casing of pep8 is counter productive and could as well be all deleted..

Patch-set: 1