Update patch set 1
Patch Set 1: Looks good to me (core reviewer)

I don't like copied code and I always wanted to be able to use first in OpenStack, so I +2.

I'm not sure it's packaged for Debian/Ubuntu yet (or I missed it), but I can make that happens in a snap.

Patch-set: 1
Label: Code-Review=+2