Update patch set 2
Patch Set 2: I would prefer that you didn't merge this

The new requirement has a lower bound, but no upper bound. Should it be capped to avoid issues from API changes? How common are those with this library?

There's no home page listed for the package, so although it mentions python 3 support in the trove classifiers, it's not clear whether that is 3.1, 3.2, or 3.3. Do you know?

Patch-set: 2
Label: Code-Review=-1