Update patch set 2
Patch Set 2:

Updating kombu seemed like it might require more testing and verification. I was trying to fix the fact that you can't install ceilometer with devstack, so I was looking for a shorter term fix. I can change approaches, though.

Which version of kombu is working for you?

Patch-set: 2