Update patch set 1
Patch Set 1:

The requirement was in tox.ini to avoid having it installed for tests, since it's really only needed for documentation. If we're going to list it in test-requires, then it no longer needs to be listed explicitly as a requirement in tox.ini because the docs environment uses the test-requires file.

Patch-set: 1