Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

(2 inline comments)

My mistake, I thought this was for updating the requirements repo from the project files. 

This approach seems unnecessarily complex. Wouldn't it be simpler to handle each line in the project requirement file as it is read? Copy blank lines and comments directly to the output and update any real requirements with the new versions before printing those. Is there a problem with that approachthat I'm not seeing?

Patch-set: 1
Label: Code-Review=-1