Update patch set 1
Patch Set 1:

> This approach seems unnecessarily complex. Wouldn't it be simpler to handle each line in the project requirement file as it is read?

Thats possible if we want to abandon the possibility to lower-case (thereby deduplify) and sort the requirements file alphabetically. The current update.py script did that. 

I already removed the alphabetical reordering in this patch, although it is debateable if we shouldn't keep it. The patch originally evolved by keeping the comments, but still do the sorting. Later I decided against keeping the re-sorting.

Patch-set: 1