Update patch set 1
Patch Set 1:

Can't the order of installation change the version of a second-level dependency that gets brought in? It seems like we've had issues in the past where there were version conflicts, but if we list the "right" version of a second-order dependency before the thing that depends on it, the conflict is resolved. Sorting would break that.

Patch-set: 1