Add HTTPretty to test requirements.
HTTPretty is a means of stubbing HTTP calls that is library independent.
It is MIT licensed and can be found: https://github.com/gabrielfalcao/HTTPretty
for examples of its use. Development is active with the 0.6.2 version
released only a week ago.

I think an approach like this is simpler than the current means of
stubbing requests or HTTPlib, more explicit in what sites requests are
being made to and it will allow a common test layout/format between
httplib, requests or whatever other http library is being used.

Change-Id: I01f291d324fd9902a4e728775019fe3c43bcf61e