Update patch set 1
Patch Set 1:

But if not all projects migrate at the same time, then this would make them clash immediately without a way to remain on python-keystoneclient 0.2.x to resolve the incompatibiliies. 

I'll be happy to submit a followup commit that raises the lower bound to 0.3 once all projects went with 0.3.x.

Patch-set: 1
Label: Code-Review=0