Clean up pass-through logic in update.py
In our gating logic (project-requirements-change.py), we ignore
tarballs, editable reqs and --find-links lines. Similarly, in update.py
we just want to pass these through on unmodified.

Add a _pass_through() function which encapsulates this logic and use it
to ignore lines when parsing the source requirements.txt and also to
pass-through matching lines from the destination requirements.txt.

Change-Id: I7c34dd65ecd866e923c117686d10e86177bb320e