Update patch set 2
Patch Set 2:

An upper cap on the next major version is a "theoretical" limitation anyway, since the intent is that we never do incompatible changes, with other words there is no need for a major version bump. 

However, *if* it happens,then fixing all modules and all maintained stable/ branches in *one* go to unbreak gating is hardly possible. I'd therefore prefer to have the upper limitation on the *major* version only. 

There is "maintenance" involved with those upper caps anyway, as they're not planned to be reached at any time.

Patch-set: 2