Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

Hi,

Well, opposite way. We must allow python-requests 1.2.3 to be used, since that's what is in Sid. See:

http://packages.debian.org/search?keywords=python-requests&searchon=names&suite=all&section=all

Please consider what's in the distributions before making arbitrary limits to older versions. On the packaging side, this is simply impossible to deal with.

Patch-set: 1
Label: Code-Review=-1