Update patch set 2
Patch Set 2:

Re-reading the old commits, I think now that the reasons for the capping on requests were not related to requests itself but due to side-effects in gating and other bugs. I remember one API incompatibility that we had to patch, but that was before 1.0. 

so I'm fine with uncapping requests.

Patch-set: 2