Update build depends to versions that grow bunnies
pip 1.4 is out. The next version of pbr will depend on it. Addiitonally,
the next version of pbr drops the use of d2to1, so let's go ahead and
remove it here. And, the next release of pbr is going to be a 1.0
release. It turns out it's in production and stuff now, let's not stick
around on the 0.x for forever.

Change-Id: I9572fce2f5f3a95e234a66cbf82ea96ee2949865