Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Supports python 2.6, 2.7, 3.2, and 3.3. Most of the project is MIT licensed with a portion licensed under what appears to be a BSD license. Also, cffi is already a transitive dependency for xattr. I do note there is a msvc dir with a .obj file in it that may be an issue for the same reason selenium is problematic with Debian. I don't know enough of the selenium situation to say for sure though.

Patch-set: 1
Reviewer: Gerrit User 4146 <4146@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1
Label: Workflow=0