Update patch set 1
Patch Set 1:

I just spoke with the neutron IRC channel and looked at the relevant code. It's somewhat non-overlapping with what we need to do, isn't predicated on a new library (it shells out to relevant commands), and is linux specific. So I don't think this is blocking issue.

Patch-set: 1