(WIP) add update.py testing to requirements
This is an attempt to add update.py testing to requirements
repository. It does so by creating a temp directory where it
puts update.py and relevant sample requirements files, runs
update.py, and looks at the resultant files.

This seemed like a sane prereq for trying to make update.py
handle oslo lines correctly, as there would be some subtlety in
that.

Because requirements repo uses requirements.txt not to specify
it's own requirements, but to specify our global list, you
currently have to delete requirements.txt and test-requirements.txt
from your directory before you can run tox. That should be fixed
in a new version of pbr.

Change-Id: Ib9b86ade4cb8317509e218aec31f32e5d08f4035