Update patch set 3
Patch Set 3:

The first error at least is:

    2013-08-06 20:28:03.477 | 2013-08-06 20:28:03   File "/usr/lib/python2.7/dist-packages/migrate/versioning/schema.py", line 10, in <module>
    2013-08-06 20:28:03.477 | 2013-08-06 20:28:03     from sqlalchemy import exceptions as sa_exceptions
    2013-08-06 20:28:03.478 | 2013-08-06 20:28:03 ImportError: cannot import name exceptions

So that is already handled in the distros by a trivial patch:
http://pkgs.fedoraproject.org/cgit/python-migrate.git/commit/?h=f19&id=603ed1d1

Now python-migrate is EOL upstream, transferring to alembic.
So as a small adjustment to relieve this blockage, it might
be appropriate to carry this simple patch within openstack?

I'm not sure how best to do that though.
Perhaps augmenting InstallVenv.post_process() in oslo-common and then propogating the patch to all projects,
and at that stage we could do the bump here.

Maybe there is a more central place we could apply patches, pbr?

Patch-set: 3
Reviewer: Gerrit User 1812 <1812@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0