add option for adding a suffix to output files
When we run update.py automatically in the devstack gate, we
have the issue that we are programatically changing files which
are actually git versioned. This makes lots of people sad pandas.

As this is all being done automatically we could actually update
the files with a suffix, then have devstack use that instead.
This will make developers happy, as they won't have to keep
resetting their devstack trees if they do development straight in
them.

Change-Id: I3d264f15ab7123f258f198b375b9667895c4104c