Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

The current version of pyghmi on pypi is 7e4f5a3. This is what pbr gives you on a tagless repo. Pypi should be updated so that a proper version is uploaded. The existing version will probably need to be removed as well to prevent confusion in pip and friends.

Also, we should make an effort to rename the project in stackforge so that our publishing tools and other jenkins stuff runs properly.

Patch-set: 1
Reviewer: Gerrit User 4146 <4146@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1
Label: Workflow=0