Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve

it is really a nightmare to upgrade pyflakes version when multiple projects are involved if all of them have a hard == on a specific version. 

I would really prefer to go with a range and then, when all projects got updated, remove the old version from the range.

Patch-set: 2