Update patch set 2
Patch Set 2:

As I know, only keystone and neutron make use of WebTest for testing, both of them can pass through the testing when setting "WebTest" in test-requirements ( then WebTest 2.0.7 is installed). If there is any other project which use WebTest, we can test/verity it. 

If all of these kind of projects can pass through the test, the new version of WebTest should not be a concern.

Patch-set: 2