Update patch set 1
Patch Set 1:

As I recall, the incubation vote was taken under the understanding that we would fully evaluate Pecan, but there was no hard graduation requirement that we must use it. We are striving for a pragmatic approach here; generally speaking, I don't think it would be prudent to force the project to use Pecan without it being fully vetted by the team.

See also: https://blueprints.launchpad.net/marconi/+spec/pecan-framework

We can discuss this further in #openstack-marconi if you like.

Patch-set: 1