cap cliff to < 1.4.5
cliff 1.4.5 bumped it's requirement of pyparsing to > 2 in a
patch release. This wedge the gate.

Uncapping pyparsing is probably the right long term thing to do,
however the infra mirror takes time to bring in new requirements.
So be more expedited and use cliff < 1.4.5 instead.

Gratuitous change to takeit out of the gate

Change-Id: Ifd799bfd99862753bd6e3cb1745bd6dc7e329f7f