Update patch set 1
Patch Set 1:

Monty and Julien,

Thank you for the quick responses.  This doesn't have to be in Havana so I can abandon the patch for now and try again once Icehouse opens up.

I am continuing to work with the developer to get the package properly uploaded to Pypi.

It is unfortunate that the package is old.  It is my understanding that not much, if anything, has changed in CIM protocols since it was written.  I think that is why it has not changed.  I will continue to investigate.

Patch-set: 1
Reviewer: Gerrit User 7198 <7198@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0