Update patch set 1
Patch Set 1:

with a last upload 2 years ago for PyCIM, I doubt it.

The important thing is that these requests come in during an open release cycle (which we aren't in), and that they are something that's maintained in pypi, and that the distros are comfortable with adding.

Patch-set: 1
Label: Code-Review=0
Label: Workflow=0