Update patch set 1
Patch Set 1:

Sean,

So, how are these types of situations most often handled?  We have a driver that requires a package that is old.  It is left to the customer to pull the package in if they are using that driver?  Do I have other options in this case?

Patch-set: 1