Update patch set 1
Patch Set 1:

Clark, ok, I get that.  To give some more detail, the problem I hit is we have an internal pip mirror and it didn't have hacking 0.7.2 (latest it had was 0.7.0).  We have some jenkins automation that is pulling the latest code from master and running 'tox --recreate -e py27,pep' against it and would keep failing in the flake8 hacking checks.  Until I compared the versions on my local workspace against what was in our jenkins slave I didn't know that hacking was backlevel in the mirror. If the version of hacking that is needed was actually in test-requirements.txt in nova, we would have failed in a more understandable way (pip won't be able to resolve the hacking-0.7.2 package in our mirror).

Patch-set: 1
Reviewer: Gerrit User 6873 <6873@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0