Update patch set 1
Patch Set 1:

I still think we have half a solution here. This change allows the upper bound of <0.8 potentially allowing this to happen again when 0.7.3 releases and so on. Setting a lower bound fixes the problem described today but not tomorrow.

Typically ranges of versions shouldn't be problematic as long as their APIs etc don't change. But hacking is special because it is used directly to check and gate our code. Probably worth seeing that mordred and jog0 have to say about this.

Patch-set: 1