Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

So all of these bugs are pointing to a different problem: there appears to be a regression in hacking 0.7.2 around the allownoqa wrapper that made some of the checks not run.  This wasn't caught because most of the tests we have are docstring tests inside of functions.

So the proper solution for this is to: fix hacking then cut a 0.8 release of hacking.

Patch-set: 1
Label: Code-Review=-1