Update patch set 1
Patch Set 1:

Thierry, the dependency in Keystone on 0.9.6 is wrong. when you check gating logs, it uses Babel 1.3 for gating, and it is passing tests on those. when you test against 0.9.6, as we do on openSUSE side, tests are failing.

Babel 1.1 is the first version that passes gating, so it makes sense to require that. it does not work with 0.9.6 anymore. 

I think at this point in time for the release it would make sense to fix the requirements of the packages to match the real requirements the project has.

Patch-set: 1