Update patch set 2
Patch Set 2:

Actually it's the odbc/cli driver within the data server driver package that has the sql header files, and you can get the odbc/cli driver package for your platform directly from ibm's fix central, it's a much smaller package. I wonder if there is a way to get something like that setup automatically in the package itself, but it would have to account for platform/architecture and version that you need (default is probably x86_64 with the latest version, which would be 10.5 right now).

Patch-set: 2