Update patch set 1
Patch Set 1:

Well, that's the whole point of why I do not want this kind of patch to get in. If it does, then the gate will artificially be blocked with some older version, and never, we will see things evolve and be maintained for newer ones. However, distributions (like for example Debian Sid, as this is the one where I upload things in) do not stay static, and the dependencies that OpenStack relies on really do have to be updated and checked. Locking packages to a specific version is an illusion.

Yes, in the short term, you will reduce the amount of work. Yes, patches will get in more easily, and there will be less breakage. I get that. Though you *do* want to see the breakage happen, simply because they need to be addressed. Yes, it's painful. But we have no choice.

Patch-set: 1