Update patch set 1
Patch Set 1:

A couple of notes... First, I'm only doing this for existing *stable* release branches (folsom and grizzly). The corresponding branches of openstack/requirements are currently IN NO WAY accurate for those releases since they were branched from master during the havana development cycle. Second, this is an attempt to normalize a set of requirements against which we test patches for those older supported releases--because we weren't applying any requirements controls during those release cycles there are in fact conflicting requirements between some projects from that timeframe. There is unfortunately some need to adjust requirements of some projects in stable release branches to match. You should be aware that this is happening anyway when integration tests run, since one and only one version of a requirement can be installed at a time, so when projects with conflicting requirements are tested together the versions of requirements are not clearly defined and prone to a somewhat random last-installed version decision.

Keep in mind that this is purely for the sake of defining the versions of dependencies we test older stable branches against. In theory downstream distro package maintainers have already decided what versions of these requirements they are carrying back when that particular OpenStack integrated release occurred. Starting with the havana release, we have more automated mechanisms to enforce requirements consistency and propose updates to projects which fall out of compliance, so one conversation which still needs to happen is whether we apply hard caps on those requirements at havana release time or looser caps with similar requirements synchronization as applied in the master branch going forward. In the meantime, this is purely at attempt to catch the folsom and grizzly releases up to some sort of consistency.

We get to choose... do we want to test old stable releases against every new version of a dependency which comes out, or do we want timely security fixes for stable branches. It's pretty much one or the other.

Patch-set: 1