Update patch set 2
Patch Set 2:

So we don't need the very last version, I got a success run with the version 0.7.6. But the version 0.7.10 is the one that tries to get installed by default right now (since there's no boundary set on the requirements) and it fix a bunch of python 3 bugs (which is something we are moving towards to support) so I don't see why not use the last version.

What you guys think?

Patch-set: 2
Label: Code-Review=0