Update patch set 1
Patch Set 1: Looks good to me (core reviewer)

The only reason why stable branches exist in the first place is to avoid duplication of work between distros, by making everyone work on the same branches (rather than private backport branches). They do not exist to spare the distributions of any work. Nobody (i.e. no distro) is apparently interested in maintaining stable/folsom anymore. The only people that need it to work in some capacity is the vulnerability management team, because we promised some support period for security patches and we need to merge them. If we are the only ones to work on it, we pick the rules. It became our branch.

Distributions can't just drop working on a stable branch and then complain that it's not tested against their preferred set of dependencies. Nobody is looking after the periodic jobs failures. The VMT doesn't have too much time to spend on it, so we simplify the problem by freezing the dependency set as much as we can.

Patch-set: 1
Label: Code-Review=+2