Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

I do not agree by principle that we should "eliminate future surprises". To the contrary, OpenStack culture should evolve that if there's a new version of something, we should as much as possible support it, because it's going to land in distributions, like it or not.

Patch-set: 1
Label: Code-Review=-1