Update patch set 1
Patch Set 1:

You wrote: "In theory downstream distro package maintainers have already decided what versions of these requirements they are carrying back when that particular OpenStack integrated release occurred." Well, this works for stable releases of distributions, though for rolling releases (like Gentoo, or with Sid), this isn't the case.

Then you wrote: "We get to choose... do we want to test old stable releases against every new version of a dependency which comes out, or do we want timely security fixes for stable branches. It's pretty much one or the other." Well, when it comes to Debian, I'm interested of having such thing (eg: stable dependency set) for Essex, not for Folsom, and soon, not for Grizzly (when I will overwrite Grizzly with Havana in Sid and Testing), though OpenStack upstream doesn't maintain Essex, and goes to have fixed dependencies for Folsom and Grizzly, which is in my case, and in the case of Ubuntu 12.10 and 13.04, not matching what is happening. Currently, I would like that Grizzly is tested against anything that appears in Debian Sid, because this is the reality I have to address.

Patch-set: 1