Update patch set 1
Patch Set 1:

change If388d872286eca15bf24a582db1bac22224a8b24 suggests that  this change will not be required on master.
If that change isn't backported to nova grizzly then since the above change suggests that boto is not going to change wrt in future, we should cap rather than exclude. I.E. we should do this on the _stable branches_ of this project:

  >=2.4.0,<2.13.0

Patch-set: 1
Label: Code-Review=0