Update patch set 1
Patch Set 1:

This looks good to me, but I think the blueprint needs to be changed to address how this will actually be used.  There aren't any changes needed in Oslo if we're going to use a third-party library.  This change shouldn't be approved without an accepted blueprint (and as I noted, accepting the Oslo one doesn't make sense now).

Patch-set: 1
Label: Code-Review=0