Update patch set 1
Patch Set 1:

Just spotted Monty's reply on the mailing list:

"Debian/Ubuntu are not packaging Icehouse for their 2.6-based releases."

This is the piece of info I needed to get on the same page -- thanks.

We'll move to both tweak our glance patch and play nicely with pbr so its not a 2.7 requirement.

Thanks!

Patch-set: 1