Update patch set 2
Patch Set 2:

do we actually know that 0.1.7 will be fixed? or is this going to need to be a hot fix again when 0.1.7 gets released?

If we don't know that we should just set this to ==0.1.4 with a comment that 0.1.5 and 0.1.6 are bad.

Patch-set: 2
Reviewer: Gerrit User 2750 <2750@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0
Label: Workflow=0