Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

This requirement spec is getting a little silly. Do we know for sure that the new release works for us? If not, can we restore the cap to 0.1.4 until we do have a release that is known to work?

Patch-set: 1