require iso8601>=0.1.8
0.1.8 supports a superset of the date formats supported in 0.1.4.
While in general this is great, it means that if we allow both, certain
usages will fail unpredictably. However, implementing either a backwards
compatible restriction on top of 0.1.8 or a forward-compatible extension
to 0.1.4 is ungainly and essentially reimplements the library. Hence, a
better course is to just advance the requirement to ensure uniformity.

Change-Id: I8fa317b30c7fb0cef83762da71f3bfc74c8847ae