Update patch set 3
Patch Set 3:

I am not convinced that the speed up introduced by storing the property value in the parent object's __dict__ is enough justification for bringing all of werkzeug in as a dependency. The library includes far far more code that we would not be using, and that may be incompatible with other WSGI code we *do* want to use.

Patch-set: 3