Update patch set 1
Patch Set 1:

@Doug

No, they don't. The test suite predates the use of TestR of even the creation of Tempest and for years we've been trying to figure out some way to sync things up with the rest of the community. However since we weren't incubated we couldn't do much other than occasionally advocate a different testing tool for projects none of us worked on.

We can't eliminate Proboscis without a major rewrite. On the plus side the current tests runs in seven seconds in "fake" mode (the Tox gates) so this isn't a huge deal. I have also discovered ways of making Proboscis work with TestR though Proboscis is still required to determine the TestSuite, though they're a bit rough around the edges, and fixing it up hasn't been a priority.

My hope is that as we start working on Tempest we will figure out how to either eliminate the use of Proboscis or bring any unique value it has to the wider community.

Patch-set: 1