Update patch set 5
Patch Set 5:

Thanks for review Julien. Let me comment on a few things. On ibm_db_sa being mocked... notice the the mock happens on ibm_db_sa.ibm_db which is an internal object referenced by ibm_db_sa module, the actual logic to determine if Exception is due to disconnect Errors is being executed with no mocks. Also, regardless of testing using the module, there is an indirect dependency on ibm_db_sa that I think is needed on openstack system the minute you make you connection string use the ibm_db_sa connection string in your sqlalchemy dialect. I don't think is a hard requirement to actually test with ibm_db_sa, but I do think it is a good test to make sure there is not regression if ibm_db_sa is modified on this area.

Patch-set: 5