Update patch set 2
Patch Set 2: I would prefer that you didn't merge this

I'm concerned that we are letting potential gate breaking situation emerge here, especially with no upper bound on the requirement. Either we need to put in an upper bound, or we need to put in some gating jobs to make sure that changes to taskflow do not break openstack projects that consume it.

Patch-set: 2
Label: Code-Review=-1
Label: Workflow=0