Update patch set 5
Patch Set 5:

Thanks for review Monty, but not sure I understand your comment... this patch has very little to do with sqlalchemy-migrate. It is in support of handling disconnects from database and letting each dialect determine if error was caused by a disconnect error. Please follow the related bug fixes waiting for this dependency.

Patch-set: 5