Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

It doesn't, as long as you use pip >= 1.4, which we should be using across the board anyway. Additionally, with 1.4, if you reference a pre-release version in the requirement, it will enable pre-release versions overall, so this could allow us to consume an alpha when we were not expecting it. (oddly)

Is this causing a problem anywhere?

Patch-set: 1
Reviewer: Gerrit User 2 <2@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1
Label: Workflow=0