Update patch set 2
Patch Set 2: I would prefer that you didn't merge this

I agree with dperaza here, except doing >=0.7.2,!=0.8.1 will just make pip use 0.8.2, which is the latest, which is what this patch already does. So as stated before, I think we should just pin to 0.7.2 for now since (a) that's the version we've been using since I can remember anyway and until (b) we can vet the 0.8.2 version against the individual projects before making a conscious (not rushed) global change.

I also don't think the sqlalchemy 0.8 support with the new migrate version holds much water as a reason to rush this in since global-requirements is still requiring sqlalchemy <= 0.7.99, so we don't get any immediate 0.8 improvements with this.

https://github.com/openstack/requirements/blob/master/global-requirements.txt#L94

Patch-set: 2