Update patch set 4
Patch Set 4:

Hi Monty-san,

Sorry but I did not mention "version" and "date" explicitly in my previous comment.

> We're well past requirements freeze - do we have any idea how this will affect the other projects and/or the distros?

When you install devstack on ubuntu it updates boto to the version which is 2.13.3. Version 2.12.0 which we would like to be the minimum version was released on Sep 4 in 2013 and so has been tested extensively since by anyone using devstack. As far as our testing has shown using the boto>=2.12.0 requirement, everything appears to be working well and we have not seen any problems. Note that some problems have been found with later versions of boto such as 2.13.0 which is already excluded from the requirements.

We'd like to push to get https://review.openstack.org/#/c/49316/ into Icehouse early in the release cycle and I think its also best to update boto early in the cycle so that it gets wider testing.

thanks,

Ozawa

Patch-set: 4