Update patch set 6
Patch Set 6:

This is really confusing.  The py3 version says it supports back to 2.5.x, so we should probably just use that, but then it goes on to say that it's not tested with Python 3, which means as far as we're concerned it doesn't work.

I'm pretty sure we're not taking new requirements that don't work with Python 3, so it seems to me that issue will need to be addressed.

Patch-set: 6