Add devstack and devstack-gate
Since projects.txt is going to be the basis for select-mirror in the
gate, we should also list projects in projects.txt that need to be
constrained by the mirror for their testing. In order to do that, we
need to not try to unconditionally update setup.py, since devstack and
devstack-gate do not have it.

Change-Id: Iea33e8f4d5453bf413f3d479b85eaed99938b302