Update patch set 6
Patch Set 6:

I did get a response from owner of ibm_db_sa and the py3 version is not something that is supported at least by IBM. The py3 package in pypi was done by another group. I explicitly says that py3 is not supported. So It think it is save to say that ibm_db_sa does not yet has support for py3. Is this a no go for dependencies? Are there any exceptions? Trying to figure out if the py3 package is a hard requirement.

Patch-set: 6