Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

I'd like to add to what Kurt said that we're indeed working on the pecan bp. Although, no code has been written yet, we've started to dig into pecan and looking for areas that we could contribute to. [0] CTRL+F Pecan

Also, we talked - several times - with Ryan, Doug and other guys behind Pecan at the summit. We agreed on contributing to Pecan and help with finding areas where it could be improved.

All that being said, blocking new versions of falcon will just hurt Marconi as a project. We also have other graduation requirements to work on and we're doing so in parallel[1].

 
[0] http://eavesdrop.openstack.org/meetings/marconi/2013/marconi.2013-12-17-15.11.log.html

[1] https://blueprints.launchpad.net/marconi/icehouse

Patch-set: 1
Label: Code-Review=+1