Update oslo.messaging requirement to use git
We probably shouldn't be gating on alphas for not-yet-released
libraries. Since oslo.messaging commits are being gated against all
projects, it should be safe to use the development version as opposed to
cutting alpha releases.

Once it'll be released, this requirement will go away and the stable
version will be added to the global-requirements.txt file.

Related-bug: #1261253

Change-Id: I98dbdccdffb7bf628cc533a1dd099505258b02f0