Update oslo.messaging requirement to latest alpha
In olso.messaging d2to1 was removed as requirement as of change
I81e3d4bf4eef833b0b55d17e2a55028972ea4b55 (which was update to
global requirements). If we update to >=1.3.0 then this change
is reflected - alleviating the need to manually install d2to1
which was causing problems for glance and tripleo as per the
related bug report.

Related-bug: #1261253

Change-Id: I98dbdccdffb7bf628cc533a1dd099505258b02f0