Add libvirt-python in prep for using pypi libvirt
Libvirt is now consumable via pypi. Add this as a requirement so that
nova can consume it in this way rather than relying on system packages.
Note that this will require newer C libvirt than is currently available
on the precise slaves. Will need working cloud archive before this can
go in.

Change-Id: I1bb37d121b8488d0d751d7f95e914e75427492b8