Update patch set 1
Patch Set 1:

Hi,

I don't want to do a +1 or a -1, just a warning here.

I'd be very careful with this, because installing the python-libvirt package makes the nova unit tests fail (which is why nova should build-conflict with python-libvirt). This is due to the fact that Nova uses its own emulation of a fake libvirt python package if it doesn't see one installed.

So I'm not sure if this has a consequence with what I just wrote above.

Patch-set: 1
Reviewer: Gerrit User 6476 <6476@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0