Require colander
Colander is a schema validation library that is going to be used in the
future as part of oslo.messaging schema validation, and probably other
part like API data type validations.

It is actively maintained by the Pylons project, and supports Python 3.
So far there's already proof of backward compatibility even if the
version is still beta.

Change-Id: I797becc37c00f45e6d878860517f42831d910fb8