Whitelist external netifaces requirement
This patch include two parts to fix the Jekins failure.

* Have tox install via setup.py develop

tox 1.6 was released, which means that we can now take advantage of
the feature we added to it - which is using setup.py develop to
install the code into the virtualenv. The logic was taken from
run_tests.sh - so the performance issues around using tox vs. using
install_venv should now be gone.

Additionally, override the tox pip install command to avoid using
the "--pre" option which is the default in tox. "--pre" means
"Include pre-release and development versions." By default, pip will
only install stable versions of software, and that is the behavior
we want.

* tox.ini(testenv.install_command): Use the --allow-external and
--allow-insecure options so that pip 1.5 and later will assent to
retrieve the netifaces package even though it's not hosted on PyPI.
The --allow-insecure option is aliased to a clearer
--allow-unverified wording in 1.5, but the old form is being used to
avoid breaking users of 1.4.x and will be valid at least through
1.6.x according to comments in the pip source.

Copied from https://review.openstack.org/#/c/65019/

Change-Id: I2af4dc6ab460fd8a368b22e3ffbde9467c23f774
Partial-Bug: #1266513