Update patch set 1
Patch Set 1:

Unfortunately no. As explained in the link provided in the commit message, functools.wrap doesn't preserve the signature (args and kwargs) of the method.

We have some argparse common code that inspects the arguments of a method to check if the required args are filled. If we add a decorator that is using functools.wrap to the decorated method, the arg parsing fails because it can't inspect correctly the method.

With @decorator.decorator, it works fine

Patch-set: 1
Reviewer: Gerrit User 9537 <9537@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0