Update patch set 3
Patch Set 3:

this package brings in the PULP LP modeling framework - http://www.coin-or.org/PuLP/ 
which works with several opensource solvers (CBC, GLPK, CPLEX, Gurobi etc)  
However the package coinor.pulp autamatically enables one of the solvers CBC. 

The pulp dependency it brings is this - https://pypi.python.org/pypi/PuLP/1.5.4  with the latest version 1.5.4 that was uploaded on 2013-03-21. 

We need this Pulp modeling framework's dependency, as the reference implementation we provide for the nova constraint solver is implemented and tested using Pulp.

Patch-set: 3