Update patch set 1
Patch Set 1:

Sorry, forgot to follow-up. So the response from Philip is that he really like to be able to track download from its own site, but since OTOH he understands our problem and is honoured to see its software be used by OpenStack, he's willing to consider it. He asked for some time to think about it.

I've sent https://review.openstack.org/#/c/69743/ to make this works in the meantime.

Patch-set: 1